More camera movement with cursor and numeric keys
Some border area has been determined to start the camera displacement
(zone_for_displacement, now at 10%). It was screen size independent, so
it will adapt to every dimensions we choose. With numeric keys, we start
a smooth transition, on each player.  There are different smoothness on
the movement, but they can tested and unified.